chore(corrosion): refactor backoff connect into single fn
