+++ src/providers/corrosion.rs
@@ -13,12 +13,102 @@ type CorrosionAddrs = Vec<crate::net::EndpointAddress>;
 type HealthCheck = Arc<atomic::AtomicBool>;
 type State = Arc<crate::config::Config>;
 
+use rand::Rng as _;
 use tryhard::backoff_strategies::{BackoffStrategy as _, ExponentialBackoff};
 
 const BACKOFF_INITIAL_DELAY: Duration = Duration::from_millis(500);
 const BACKOFF_MAX_DELAY: Duration = Duration::from_secs(30);
 const BACKOFF_MAX_JITTER: Duration = Duration::from_secs(2);
 const CONNECTION_TIMEOUT: Duration = Duration::from_secs(5);
+
+/// Repeatedly attempts `connect` against all `endpoints` in parallel with
+/// backoff, resolving with the first successful connection
+///
+/// Connecting to all endpoints in parallel prevents down/slow servers in the
+/// list from unnecessarily delaying connections to healthy servers. Currently
+/// we just go with the first server that we can successfully connect to, but
+/// in the future we could connect to multiple servers simultaneously.
+async fn connect_first<T, F, Fut>(
+    endpoints: &CorrosionAddrs,
+    connect: F,
+) -> crate::Result<(T, crate::net::EndpointAddress)>
+where
+    T: Send + 'static,
+    F: Fn(crate::net::EndpointAddress) -> Fut,
+    Fut: std::future::Future<Output = crate::Result<T>> + Send + 'static,
+{
+    let mut backoff = ExponentialBackoff::new(BACKOFF_INITIAL_DELAY);
+
+    let retry_config =
+        tryhard::RetryFutureConfig::new(u32::MAX).custom_backoff(|attempt, error: &_| {
+            tracing::info!(attempt, "Retrying to connect");
+            // reset after success
+            if attempt <= 1 {
+                backoff = ExponentialBackoff::new(BACKOFF_INITIAL_DELAY);
+            }
+
+            let mut delay = backoff.delay(attempt, &error).min(BACKOFF_MAX_DELAY);
+            delay += Duration::from_millis(
+                rand::rng().random_range(0..BACKOFF_MAX_JITTER.as_millis() as _),
+            );
+
+            tracing::warn!(?error, "Unable to connect to the corrosion server");
+            tryhard::RetryPolicy::Delay(delay)
+        });
+
+    tryhard::retry_fn(|| {
+        tracing::debug!(
+            server_count = endpoints.len(),
+            "attempting to connect to corrosion servers"
+        );
+
+        let mut js = tokio::task::JoinSet::new();
+
+        for addr in endpoints.iter().cloned() {
+            let fut = connect(addr.clone());
+            js.spawn(async move { (fut.await, addr) });
+        }
+
+        let num_endpoints = endpoints.len();
+
+        async move {
+            match tokio::time::timeout(CONNECTION_TIMEOUT, async {
+                while let Some(join_result) = js.join_next().await {
+                    match join_result {
+                        Ok((result, addr)) => {
+                            match result {
+                                Ok(connection) => {
+                                    return Ok((connection, addr));
+                                }
+                                Err(error) => {
+                                    tracing::warn!(address = %addr, %error, "failed to connect");
+                                }
+                            }
+                        }
+                        Err(join_error) => {
+                            if join_error.is_panic() {
+                                tracing::error!(
+                                    ?join_error,
+                                    "panic occurred in task attempting to connect to corrosion endpoint"
+                                );
+                            }
+                        }
+                    }
+                }
+
+                eyre::bail!("no successful connections could be made to {num_endpoints} possible corrosion servers");
+            })
+            .await
+            {
+                Ok(Ok(cae)) => Ok(cae),
+                Ok(Err(err)) => Err(err),
+                Err(_) => eyre::bail!("timed out after {CONNECTION_TIMEOUT:?} attempting to connect to one of {num_endpoints} possible corrosion servers"),
+            }
+        }
+    })
+    .with_config(retry_config)
+    .await
+}
 
 #[derive(Copy, Clone, Debug)]
 pub enum CorrosionMode {
